[FIX] Move examples utilites out of subdir
